Revert "Fix clamping of numbers at the upper bounds of u64 and u128 by implementing a Uint64 type"
This reverts commit 18dc723f45843b816321fb1f184ef07a806d6d7b.

+++ serde-generate/src/dart.rs
@@ -289,7 +289,7 @@ where
             U8 => "int".into(),
             U16 => "int".into(),
             U32 => "int".into(),
-            U64 => "Uint64".into(),
+            U64 => "int".into(),
             U128 => "Uint128".into(),
             F32 => "double".into(),
             F64 => "double".into(),
